tests/submission/Microsoft/apis/getElementsByClassNameFrame_002.htm

<!DOCTYPE html>
<html><head>
        <title>getElementsByClassName</title>
        <meta content="get class from frameset" name="description">    
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
        <div id="log"></div>
        <body>bar
        <script type="text/javascript">
            test(function()
            {
                var collection = parent.document.getElementsByClassName("foo");
                assert_equals(collection.length, 3);
                assert_equals(collection[0].parentNode.nodeName, "HTML");
                assert_equals(collection[1].parentNode.nodeName, "FRAMESET");
                assert_equals(collection[2].parentNode.nodeName, "FRAMESET");
            }, "get class from frameset");
        </script>
    
       
</body></html>
